requirements.txt
tensorflow
tensorflow-hub
tensorflow-text
streamlit
diffusers
transformers
torch
Pillow
scipy
